+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v3.6.0] - 2021-11-08
+### Added
+- Parameter `log_config_sample_rate`
+### Changed
+- `example` folder contents updated to Terraform 0.15+ format
+
 ## [v3.5.0] - 2021-07-27
 ### Added
 - Add custom health check ports into firewall allow rule